agent/requirements.txt
openai>=1.30.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
beautifulsoup4>=4.12.0
duckduckgo-search>=7.0.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
